Move CSS and JS to roof of the HTML
We missed that each model would have a separate CSS and JS. This is
unnecessary, and can slow down the development. We move CSS and JS to
the root of HTML, and share them among the models.
