+++ htmlgen/common.py
@@ -1,0 +1,6 @@
+"""Provide common elements used between the generators."""
+
+I = "  "
+II = I * 2
+III = I * 3
+IIII = I * 4
