+++ htmlgen/main.py
@@ -7,14 +7,16 @@ import pathlib
 import sys
 from typing import Tuple, Optional, List
 
-import asttokens
-from aas_core_codegen import intermediate
-from aas_core_codegen.common import Error, LinenoColumner
-import aas_core_codegen.run
 import aas_core_codegen.parse
+import aas_core_codegen.run
+import asttokens
+import pygments.formatters
+from aas_core_codegen import intermediate
+from aas_core_codegen.common import Error
 from icontract import require, ensure
 
 import htmlgen.for_metamodel
+from htmlgen.common import I
 
 
 @require(lambda model_path: model_path.exists() and model_path.is_file())
@@ -80,6 +82,99 @@ def _load_meta_model(
     assert ir_symbol_table is not None
 
     return (ir_symbol_table, atok), None
+
+
+# fmt: off
+@ensure(
+    lambda result:
+    not result.startswith('\n')
+    and not result.startswith(' ')
+    and not result.startswith('\t'),
+    "No prefix whitespace"
+)
+@ensure(
+    lambda result:
+    result.endswith('\n'),
+    "Trailing new line is mandatory"
+)
+# fmt: on
+def _generate_css() -> str:
+    """Generate the CSS for the whole docs."""
+    pygments_style_def = pygments.formatters.HtmlFormatter().get_style_defs(
+        ".highlight"
+    )
+
+    return f"""\
+a.aas-anchor-link {{
+{I}color: blue;
+{I}color: rgba(0, 0, 0, 0.2);
+{I}font-size: xx-small;
+{I}vertical-align: super;
+{I}text-decoration: none;
+}}
+
+div#menu {{
+{I}min-height: 95vh;
+{I}height: 95vh;
+}}
+
+div#content {{
+{I}min-height: 95vh;
+{I}height: 95vh;
+}}
+
+footer#footer {{
+{I}min-height: 5vh;
+{I}height: 5vh;
+}}
+
+.aas-type-annotation {{
+{I}font-family: monospace;
+}}
+
+dd {{
+{I}padding-left: 2em;
+}}
+
+.highlight pre {{
+{I}white-space: pre;
+{I}border-radius: inherit;
+{I}display: inherit;
+{I}background-color: inherit;
+{I}border: inherit;
+{I}color: inherit;
+{I}padding: 0.5em;
+}}
+{pygments_style_def}
+"""
+
+
+# fmt: off
+@ensure(
+    lambda result:
+    not result.startswith('\n')
+    and not result.startswith(' ')
+    and not result.startswith('\t'),
+    "No prefix whitespace"
+)
+@ensure(
+    lambda result:
+    result.endswith('\n'),
+    "Trailing new line is mandatory"
+)
+# fmt: on
+def _generate_js() -> str:
+    """Generate the JavaScript for the whole docs."""
+    return f"""\
+var activeElement = document.querySelector("#menu .active");
+if (activeElement) {{
+{I}var rect = activeElement.getBoundingClientRect();
+{I}var menu = document.getElementById("menu");
+{I}var fontSize = parseFloat(getComputedStyle(menu).fontSize);
+{I}// Go two lines up so that the reader immediately sees that there are more options.
+{I}menu.scrollTop = rect.top - 2 * fontSize;
+}}
+"""
 
 
 def main() -> int:
@@ -152,7 +247,9 @@ We generated the HTML documentation for the following versions of the meta-model
 """,
         encoding="utf-8",
     )
-    # TODO (mristin, 2023-01-13): generate the index page
+
+    (html_dir / "base.css").write_text(_generate_css(), encoding="utf-8")
+    (html_dir / "atTheEndOfBody.js").write_text(_generate_js(), encoding="utf-8")
 
     return 0
